refactor: changed all the names from Brawle to Brawdle #19
