+++ Proyect/Brawle/src/main/java/es/joja/Brawdle/BrawdleApplication.java
@@ -1,13 +1,13 @@
-package es.joja.Brawle;
+package es.joja.Brawdle;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class BrawleApplication {
+public class BrawdleApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(BrawleApplication.class, args);
+		SpringApplication.run(BrawdleApplication.class, args);
 	}
 
 }
