+++ Proyect/Brawle/src/main/java/es/joja/Brawdle/contract/RolesContract.java
@@ -1,4 +1,4 @@
-package es.joja.Brawle.contract;
+package es.joja.Brawdle.contract;
 
 public class RolesContract {
     private RolesContract() {}
