+++ Proyect/Brawle/src/main/java/es/joja/Brawdle/contract/UserContract.java
@@ -1,4 +1,4 @@
-package es.joja.Brawle.contract;
+package es.joja.Brawdle.contract;
 
 public class UserContract {
     private UserContract() {}
