+++ Proyect/Brawle/src/main/java/es/joja/Brawdle/controller/AppController.java
@@ -1,0 +1,4 @@
+package es.joja.Brawdle.controller;
+
+public class AppController {
+}
