+++ Proyect/Brawle/src/main/java/es/joja/Brawdle/dao/DDBBConnManager.java
@@ -1,4 +1,4 @@
-package es.joja.Brawle.dao;
+package es.joja.Brawdle.dao;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
