+++ Proyect/Brawle/src/main/java/es/joja/Brawdle/dao/LegendDAO.java
@@ -1,4 +1,4 @@
-package es.joja.Brawle.dao;
+package es.joja.Brawdle.dao;
 
 import java.util.ArrayList;
