--- Proyect/Brawle/src/main/java/es/joja/Brawle/controller/AppController.java
@@ -1,4 +1,0 @@
-package es.joja.Brawle.controller;
-
-public class AppController {
-}
